MNM-40-refactor: useImageUpload 파일 위치 변경
